Split travel restrictions out to TravelRestrictionsAPI, to better match our diagrams.
